+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next release
 
 - fix: change dep of rustdoc on push
+- feat: modify the hash_bytes functions in `poseidon` and `pedersen` for dynamic
+  data length
 - chore: add real class hash values for genesis config
 - feat: unification of the DA interface
 - feat: use resolver 2 for cargo in the workspace